+++ src/hal/indi/mod.rs
@@ -1,4 +1,4 @@
-mod sexagesimal;
+mod num_format;
 mod connection;
 mod drivers;
 mod base64;
@@ -6,6 +6,6 @@ mod xml_reader;
 mod xml_helper;
 mod error;
 
-pub use sexagesimal::*;
+pub use num_format::*;
 pub use connection::*;
 pub use drivers::*;
